Progress on multi-grid generation
This is a lot more convoluted than I expected it to be. North-south generation now works on a straight road. Need to implement fixes for south-facing roads which then turn right (west) as currently a skyscraper may generate and block them. Road turning should happen before skyscraper generation and cancel skyscraper generation if appropriate - and if a corner is set to generate, flag itself as such and just keep placing straights until it can place a corner.