Complete overhaul of tile registry system
No more EnumTile!
No more Tile Registration!
Tiles are registered in one place in code, passed to the array in the registry prefab, and must have ID set on the tile prefab
Just generally a million times nicer and easier to add new tiles in the future.

Also started work on a neighbour awareness system for tiles. Roads now build walls when at the map edge or next to air.